invite_project/requirements.txt
asgiref==3.7.2
dj-database-url==2.1.0
Django==4.2.4
djangorestframework==3.14.0
drf-yasg==1.21.7
gunicorn==21.2.0
inflection==0.5.1
packaging==23.1
psycopg2==2.9.7
pytz==2023.3
PyYAML==6.0.1
sqlparse==0.4.4
typing_extensions==4.7.1
uritemplate==4.1.1
